fix: no longer transform import before sending it to the backend (#1066)
Previously, the frontend would convert the JSON import to an instance before sending it to the backend.
This, however, lead to the backend rejecting the import as an instance is only meant to be used during validation.
Now, only the backend validates the import, with the frontend just passing on the JSON.

This problem was introduced in #1055 as before that the conversion to an instance was faulty in a way
that the backend still understood and accepted the imports.

Fixes #1065

Signed-off-by: ClFeSc <68013019+ClFeSc@users.noreply.github.com>